Modifications to Database MySQL file
Changed Date and Time format from DD/MM/YYYY and XX:XX[AM/PM]
to YYYY-MM-DD and HH:MM:SS.
This is to conform with MySQL's datatype DATE and TIME.

Database Modifications:
    Added table passengerlist, which has the FlightID, UserID, and Seat number.
    Modified table crews, should be finished and working.
    Modified table flights to work with MySQL DATE and TIME.
    Modified table users, added columns SignUpDate and CreditCardNumber.
    Modified table airpots, added city column

Created 3 MySQL files, the first (Initialize.sql) that stores the
queries for inserting data into the tables.

The second (EventTriggers.sql) stores the triggers for events that
might occur to the database and tables. Also added some comments on some
triggers that need to be implemented.

The third (Queries.sql) has some basic queries that can be run to get the
data in some of the tables.