Added some Java files and modified database Changed database name from FlightSystemDatabase to FlightSystem.
Minor changes to database, extra information here and there.

Created Java project.
Started work on database connection and object creation.
Created Java files for some of the objects needed for the program.
    Airport, Flight, Plane, Staff, User, PassengerList

Created DataImport.java to house the connection to the database.
It will also house the functions that actually make queries to the database.